Added page tracking for the toolbox
